Split out content encoding enum
