+++ tower-http/src/compression_utils.rs
@@ -14,7 +14,7 @@ use std::{
 use tokio::io::AsyncRead;
 use tokio_util::io::{poll_read_buf, StreamReader};
 
-use crate::BodyOrIoError;
+use crate::{content_encoding::SupportedEncodings, BodyOrIoError};
 
 #[derive(Debug, Clone, Copy)]
 pub(crate) struct AcceptEncoding {
@@ -25,7 +25,7 @@ pub(crate) struct AcceptEncoding {
 
 impl AcceptEncoding {
     #[allow(dead_code)]
-    pub(crate) fn to_header_value(&self) -> Option<HeaderValue> {
+    pub(crate) fn to_header_value(self) -> Option<HeaderValue> {
         let accept = match (self.gzip(), self.deflate(), self.br()) {
             (true, true, true) => "gzip,deflate,br",
             (true, true, false) => "gzip,deflate",
@@ -37,42 +37,6 @@ impl AcceptEncoding {
             (false, false, false) => return None,
         };
         Some(HeaderValue::from_static(accept))
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn gzip(&self) -> bool {
-        #[cfg(any(feature = "decompression-gzip", feature = "compression-gzip"))]
-        {
-            self.gzip
-        }
-        #[cfg(not(any(feature = "decompression-gzip", feature = "compression-gzip")))]
-        {
-            false
-        }
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn deflate(&self) -> bool {
-        #[cfg(any(feature = "decompression-deflate", feature = "compression-deflate"))]
-        {
-            self.deflate
-        }
-        #[cfg(not(any(feature = "decompression-deflate", feature = "compression-deflate")))]
-        {
-            false
-        }
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn br(&self) -> bool {
-        #[cfg(any(feature = "decompression-br", feature = "compression-br"))]
-        {
-            self.br
-        }
-        #[cfg(not(any(feature = "decompression-br", feature = "compression-br")))]
-        {
-            false
-        }
     }
 
     #[allow(dead_code)]
@@ -88,6 +52,44 @@ impl AcceptEncoding {
     #[allow(dead_code)]
     pub(crate) fn set_br(&mut self, enable: bool) {
         self.br = enable;
+    }
+}
+
+impl SupportedEncodings for AcceptEncoding {
+    #[allow(dead_code)]
+    fn gzip(&self) -> bool {
+        #[cfg(any(feature = "decompression-gzip", feature = "compression-gzip"))]
+        {
+            self.gzip
+        }
+        #[cfg(not(any(feature = "decompression-gzip", feature = "compression-gzip")))]
+        {
+            false
+        }
+    }
+
+    #[allow(dead_code)]
+    fn deflate(&self) -> bool {
+        #[cfg(any(feature = "decompression-deflate", feature = "compression-deflate"))]
+        {
+            self.deflate
+        }
+        #[cfg(not(any(feature = "decompression-deflate", feature = "compression-deflate")))]
+        {
+            false
+        }
+    }
+
+    #[allow(dead_code)]
+    fn br(&self) -> bool {
+        #[cfg(any(feature = "decompression-br", feature = "compression-br"))]
+        {
+            self.br
+        }
+        #[cfg(not(any(feature = "decompression-br", feature = "compression-br")))]
+        {
+            false
+        }
     }
 }
